Add a separate config file to use for rpm installs
In RHEL/Fedora systems we want to use a slightly different set
of defaults, but we do not want to change the standard config file
because there are many assumptions about its configuration in
openssl upstream tests.

So we create a separate one to use to override the default on on
installation.

This config file differs from upstream for:
- CA directory tree paths
- Instructions about legacy provider
- Default certificate digest (set to sha256)

Signed-off-by: Simo Sorce <simo@redhat.com>